% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mineral}
\alias{mineral}
\title{Mineral: grades}
\format{
A data frame with 4 columns:
\describe{
\item{Name}{Name of the phosphate mining and beneficiation complex}
\item{Confidence}{na}
\item{Value}{Grade, \% P2O5}
\item{Source}{Identifier for document(s) used as data source(s) (see \code{\link{sources}} for details)}
}
}
\usage{
mineral
}
\description{
Manually collected complex-specific data on the average grades of 
representative phosphate-bearing mineral associated with ore deposit.
The grade refers to the concentration of P2O5 mineral contains.
All data points are provided on the annual basis and linked to their respective source document(s).
}
\details{
Minerals containing phosphates in their chemical composition are classified as phosphate minerals.
About 200 of them are known to contain 1\% or more P2O5. 95\% of P in
the Earth's crust is bound in the various forms of apatites (Krauss et al. 1984); a number of other
iron- and aluminum-rich phosphate minerals, such as crandallite, millisite, wavellite and strengite, can
occur in the supergene enriched zones above a primary phosphate deposit (Pufahl and Groat 2016).
In igneous deposits, apatites approaching composition of pure fluorapatite are most common (McClellan and van Kauwenbergh 1990).
Most of them contain the carbonate fluorapatite variety assigned to francolite (McClellan and van Kauwenbergh 1990) - carbonate
fluorapatites containing more than 1\% fluorine and appreciable amounts of CO2.
See \code{\link{sources}} for classification of representative complex-specific phosphate-bearing minerals used in this dataset.
}
\references{
\itemize{
  \item Krauss, U., Saam, H., & Schmidt, H. (1984). International strategic minerals inventory. Summary report - Phosphate. USGS Circular 930-C.
  \item McClellan, G. H. (1980). Mineralogy of carbonate fluorapatites. Journal of the Geological Society, 137: 675-681. https://doi.org/10.1144/gsjgs.137.6.0675
	 \item McClellan, G. H., & van Kauwenbergh, S. J. (1990). Mineralogy of sedimentary apatites. Journal of the Geological Society, 137(6): 675-681. https://doi.org/10.1144/gsjgs.137.6.0675
  \item Pufahl, P. K., & Groat, L. A. (2016). Sedimentary and igneous phosphate deposits: formation and exploration: an invited paper. Economic Geology, 112(3): 483-516. https://doi.org/10.2113/econgeo.112.3.483
}
}
\keyword{datasets}
