% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mining_complexes}
\alias{mining_complexes}
\title{General properties of phosphate mining and beneficiation complexes}
\format{
A data frame with 8 columns:
\describe{
\item{Name}{Name of mining and beneficiation complex}
\item{Region}{Region name (according to the IFA classification)}
\item{Country}{Country name (according to the U.S. Geological Survey)}
\item{Subnational_Level}{Sub-national administrative area}
\item{Company}{Name of the company owning mine(s) and beneficiation plant(s)}
\item{Rock_type}{Type of rock formation: sedimentary or igneous}
\item{Status}{Status of mining project: operational or in development}
\item{Capacity}{Capacity in million tonnes of mass of phosphate rock (PR)}
}
}
\usage{
mining_complexes
}
\description{
'\emph{Phosphate mining and beneficiation complex}': a phosphate ore deposit with mine(s) and beneficiation plant(s) designated by name with definable average chemical composition of ore
}
\keyword{datasets}
