% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{phosphateRock}
\alias{phosphateRock}
\title{Global dataset on phosphate mining and beneficiation}
\description{
The dataset covers worldwide production of phosphate rock (PR) at the level 
of phosphate mining and beneficiation complex. The objective is to gather 
complex-specific data on the P content of mined and beneficiated resource and 
on the recovery rates of beneficiation process. Phosphate mining and 
beneficiation complex refers to a phosphate ore deposit with adjacent mine(s) 
and beneficiation plant(s).
}
\seealso{
\href{../doc/Introduction.pdf}{phosphateRock::Introduction}
}
