% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PR}
\alias{PR}
\title{Phosphate rock (PR): grades}
\format{
A data frame with 4 columns:
\describe{
\item{Name}{Name of the phosphate mining and beneficiation complex}
\item{Confidence}{High (relevant company's data), medium or low}
\item{Value}{Grade, \% P2O5}
\item{Source}{Identifier for document(s) used as data source(s) (see \code{\link{sources}} for details)}
}
}
\usage{
PR
}
\description{
Manually collected complex-specific data on the average grades of PR produced.
The grade refers to the concentration of P2O5 PR contains.
All data points are provided on the annual basis and linked to their respective source document(s).
}
\details{
PR is considered as run of beneficiation plant - ore after beneficiation averaged over processing streams.
}
\keyword{datasets}
