% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recovery_mass}
\alias{recovery_mass}
\title{Beneficiation: mass recoveries}
\format{
A data frame with 5 columns:
\describe{
\item{Name}{Name of the phosphate mining and beneficiation complex}
\item{Confidence}{\emph{High} (relevant company's data), \emph{Medium} or \emph{Low}}
\item{Value}{PR-to-ore ratio}
\item{Source}{Identifier for document(s) used as data source(s) (see \code{\link{sources}} for details)}
}
}
\usage{
recovery_mass
}
\description{
Manually collected complex-specific data on the average mass recovery rates.
The mass recovery refers to the ratio of the mass of produced PR to the mass of ore prior beneficiation.
All data points are provided on the annual basis and linked to their respective source document(s).
}
\details{
A data source contains information whether primary data was used or value was
estimated from \code{\link{recovery_mineral}}.
}
\keyword{datasets}
