% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recovery_mineral}
\alias{recovery_mineral}
\title{Beneficiation: mineral recoveries}
\format{
A data frame with 5 columns:
\describe{
\item{Name}{Name of the phosphate mining and beneficiation complex}
\item{Confidence}{\emph{High} (relevant company's data), \emph{Medium} or \emph{Low}}
\item{Value}{Ratio of \% P2O5 in PR to \% P2O5 in ore}
\item{Source}{Identifier for document(s) used as data source(s) (see \code{\link{sources}} for details)}
}
}
\usage{
recovery_mineral
}
\description{
Manually collected complex-specific data on the average mineral recovery rates.
The mineral recovery refers to the percent of P2O5 recovered in the PR from the ore after beneficiation.
All data points are provided on the annual basis and linked to their respective source document(s).
}
\details{
A data source contains information whether primary data was used or value was
estimated from \code{\link{recovery_mass}}.
}
\keyword{datasets}
