% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sources}
\alias{sources}
\title{List of data sources}
\format{
A data frame with 4 columns:
\describe{
\item{Source}{Identifier for citation}
\item{Reference}{Brief bibliographical reference(s) to source document(s)}
\item{Year}{Year of relevance}
\item{Note_confidence}{Notes on quality of source data}
\item{Comment}{Additional notes on data collection}
}
}
\usage{
sources
}
\description{
General information about data sources, including their identifiers, references to source document(s),
data relevance, and notes on data collection.
}
\details{
For bibliography, see 
\href{../doc/References.pdf}{phosphateRock::References}
}
\keyword{datasets}
