% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mining_complexes}
\alias{mining_complexes}
\title{Phosphate mining and beneficiation complexes: general information}
\format{
A data frame with 8 columns:
\describe{
\item{Name}{Name of the complex}
\item{Region}{Region name (according to the IFA classification)}
\item{Country}{Country name (according to the U.S. Geological Survey)}
\item{Subnational_area}{Sub-national administrative area (e.g., region, state, province)}
\item{Company}{Name of the company owning mining and beneficiation facilities}
\item{Rock_type}{Type of rock formation: \emph{sedimentary} or \emph{igneous}}
\item{Status}{Status of mining project: \emph{operational} or \emph{in development}}
\item{Capacity}{Capacity in million tonnes of mass of phosphate rock (PR)}
}
}
\usage{
mining_complexes
}
\description{
General properties of individual phosphate mining and beneficiation complexes, 
including their name, geographical location, type of rock, ownership, 
production capacity, and operational status.
}
\keyword{datasets}
